requirements.txt
python-magic==0.4.27
pytest==8.0.0
